fix fmt

+++ pgx/mahjong/mini_mahjong.py
@@ -73,7 +73,7 @@ class Hand:
     @staticmethod
     @jit
     def can_tsumo(hand: jnp.ndarray) -> bool:
-        heads, valid = 0, 1
+        heads, valid = jnp.int32(0), jnp.int32(1)
         for suit in range(3):
             valid &= Hand.cache(
                 jax.lax.fori_loop(
